Fix same color fill bug
There was an issue where using the fill tool on the same color as the currently selected color would cause the application to crash. This commit adds a check on the fill tool to see if the cell clicked has the same color as the current color and stops the fill tool from running if true.